Apply transforms during build not during save
